fix(cli): rename knowledge.py to knowledge_cmd.py to avoid module collision
The test imports `knowledge.gardener` from the monolith backend, but
`tools/cli/knowledge.py` shadows it since both use `imports = ["."]`.
Renaming to `knowledge_cmd.py` resolves the collision.
