+++ tools/cli/knowledge_cmd.py
